Karenswry external readmes (#587)
* Initial touch up, needs further clarification

* New updated s.description. Not verified yet

* Changed needed to necessary, removed redundant words

Had a lengthy conversation with native English speaking tech writers about "needed vs necessary" recently. It's much better to use "necessary" in such situations as it is as a legit adjective and "needed" is perceived as a verb forced to act as an adjective in their eyes. That's not a set-in-stone rule, but "necessary" is just better here.

* Update themis.podspec

* Update rbthemis.gemspec

* Update rbthemis.gemspec

* Added the missing dot (in addition to description update)

* Updated the description, fixed missing dot, added Python

* Shortened and humanised a bit

* Apply suggestions from code review

rm "wrappers"

Co-authored-by: vixentael <vixentael@users.noreply.github.com>