Cleanup and test Soter KDF (#585)
* Improve error handling: soter_kdf()

I've just wondered into this file to check whether soter_kdf() does
soter_wipe() on whatever it needs to wipe. And I hit a jackpot: there
were like at least three ways to cause a segfault with this function
by accidentally passing it invalid arguments. This is unacceptable.

Equip soter_kdf() with proper parameter validation. While we're here
simplify some computations and wipe other stuff that we might leave.

* Update API documentation: soter_kdf()

Update Doxygen API comments so that they are of the same quality as
PBKDF2 documentation.

* Test vectors for soter_kdf()

I have not found publicly available test vectors for ZRTP KDF so I had
to write an independent tool to generate them. This has uncoverted the
fact that soter_kdf() does not implement ZRTP KDF as described in RFC,
but uses a slightly modified version of it. It should be fine from
security perspective as NIST SP 800-108 says it's okay to omit the
length field from the hash sum.

Verify some other features of soter_kdf() function too: skipping NULL
context buffers and usage of implicit keys.

* Clarify standards compliance

Do not claim that we implement "RFC 6189 4.5.1", just refer to it as
out inspiration or something. Comment the implementation differences.