Improve naming of new parsing utilities
Add a "stream" prefix so that they are better associated with file
streams and similar wStream utilities from WinPR.

Unify read and write API to use in/out pointers in last position, with
first argument always being the stream to modify, with its new value
returned from the function.

The behavior is the same, but the API is neater and easier to read.