Simplify auth tag length output
Encryption paths of master key and passphrase API has an unnecessary
and weird check of auth tag length output. Strictly speaking, we should
measure the size of tag but we just know that it's 12 bytes for AES-GCM.

Tweak the pointer type in internal helper function so that it is able to
directly write the auth tag length into the header struct. This way we
don't have to check the length is unchanged.

However, since Soter API outputs size_t, we do need to check that it's
safe to truncate the value returned by Soter.