Benchmark passphrase API of Secure Cell (#582)
Right now we don't have support for this API in RustThemis so we call
C code directly in decryption benchmark to prepare the input data.

Note that due to the nature of the passphrase API -- KDF makes repeated
invocations purposefully slow -- the benchmarks are abysmally slow too
and we have to cut down the sample sizes to execute them in a resonable
time frame. And we don't need to have much variety in the message size
since execution time is completely dominated by KDF costs until we get
to gigabyte-sized messages.

There is quite a few "inheritance" in the code from master key
benchmarks but I'll leave it around for now until we are sure what
utilities can be shared. Plus, Criterion has issues with reusing code
so we might need to introduce additional helper crates for that which
I would like to avoid doing.