Action: Themis Core
This is the first action we add so it sets an example. All testing
actions are triggered by either pushing to important integration
branches, or by submitting a pull request touching relevant files,
or on schedule at 06:00 UTC every day.

All of them are going to have at least "unit-tests" and "examples" jobs,
with some more as appropriate. Since Themis Core is the most diverse
thing that we have, it has quite a few jobs for checking it.

Code examples for Themis Core are self-contained tests, we just need to
build and run them. However, there were some minor issues with the code
producing warnings that we resolve for a clean build.