Action: GoThemis
This is one of the most straighforward actions -- but not so fast!
GoThemis build actually requires GOTHEMIS_IMPORT variable to be set,
just be aware of that. "make test_go" relies on it.

Take care to test against multiple versions of Go. We had issues with
that in the past so this environment is important. Unfortunately,
installing multiple versions of Go is not that easy with available
actions. But I don't want to write a new one (or some shell scripts)
so we simply run them in parallel.

Also, touch up code examples: gofmt them and make sure they os.Exit(1)
when then fail.