Action: PHPThemis
Remember that we have two separate code bases for PHP 5 and PHP 7.
Furthermore, PHP is somewhat "old school" with installation so use
a convenient PPA to install various versions in parallel.

There is some weird interference between modules so we make sure to
*not* install php-fpm which seems to break PHP distribution. Anyhow,
thank you, Ondrej, for maintaining this repository!

Our Makefile expects PHP to be available as "php" in PATH, so we use
update-alternatives to fix up the symlinks in the system.

We test examples with the latest version only to reduce the matrix.

Also note that PHPThemis currently does not support PHP 7.3+.