Action: JavaThemis
Well, actually, more like AndroidThemis right now, but the actions are
named after code bases, not platforms.

There are few nice words that I have for Android on CI. Let's just leave
it at that it's abysmally slow. Though, this script seems to do the job.
Most of the time. Sometimes we still fail to wait for emulator to boot.
If we're lucky, it boots in 3-4 minutes. If we're not lucky it gets
stuck for 15 minutes.

JavaThemis currently does not have a standlone unit-test suite so we
have to test it via instrumentation tests on Android by running the
entire emulator. However, we should still do these tests as Android
build process is tricky and sometimes it does fail.