Action: RbThemis
Well this is easy. The only hard part is installing RVM which for some
reason really does not want to play nice, arbitrarily requiring you to
relogin into your shell to start working, etc. We can't to that on CI :(

There is a nice PPA -- thanks, Rael! -- which helps a bit, but there are
still some things that we need to do manually.

Also, tweak Secure Comparator example to actually stop once the
comparison is complete, not just sit there in an infinite loop.