Run GoThemis tests across multiple Go versions
We have found out that GoThemis seems to be broken with Go 1.11.
Currently we test it versus whatever version is in the image. Instead,
download latest supported versions of Go and test against them.

We would like to make use of CircleCI caches -- because that's around
500 MB that we're downloading -- but it turns out that it does not
provide much of a speedup, so we simply pull tarballs from Google.
Ideally, they should be a part of Docker image we use.