Regression test for bytes.Buffer bug in go 1.11
Add a test which expemplifies the bug in CGo 1.11 and earlier versions
(at least since 1.6). It has been fixed in 1.12 but we still have to
support some earlier versions.

The bug causes the following test to panic. For some unknown reason
it seems that only Secure Cell is susceptible to it, and only with
"context" argument being used with bytes.Buffer vOv

The panic itself is caused by a bug in CGo implementation which triggers
a false positive in a pointer check, which erroenously verifies not only
the context []byte buffer, but the entire bytes.Buffer object. Read more
in the linked GitHub issue if you are interested.