Test GoThemis with Go 1.11~1.14 on CircleCI (#595)
* Run GoThemis tests across multiple Go versions

We have found out that GoThemis seems to be broken with Go 1.11.
Currently we test it versus whatever version is in the image. Instead,
download latest supported versions of Go and test against them.

We would like to make use of CircleCI caches -- because that's around
500 MB that we're downloading -- but it turns out that it does not
provide much of a speedup, so we simply pull tarballs from Google.
Ideally, they should be a part of Docker image we use.

* Regression test for bytes.Buffer bug in go 1.11

Add a test which exemplifies the bug in CGo 1.11 and earlier versions
(at least since 1.6). It has been fixed in 1.12 but we still have to
support some earlier versions.

The bug causes the following test to panic. For some unknown reason
it seems that only Secure Cell is susceptible to it, and only with
"context" argument being used with bytes.Buffer vOv

The panic itself is caused by a bug in CGo implementation which triggers
a false positive in a pointer check, which erroneously verifies not only
the context []byte buffer, but the entire bytes.Buffer object. Read more
in the linked GitHub issue if you are interested.

* Update regression test

After discussion we have decided to not include a fix for this behavior
because of miscellaneous complications. However, the test case is still
there. Just update it so that it fails when our code unexpectedly panics
but passes if the panic is raised for Go 1.11 and below.

* Simple benchmarks for Secure Message

We've got curious about impact of copying on Secure Message performance
so here are some primitive benchmarks to measure the impact (if any).