Exercise fuzzers during CI builds (#579)
* Improve AFL composability

Instead of using our special AFL_CFLAGS and AFL_LDFLAGS, just set them
per-target as we do in other places. We'd like to build AFL fuzzers
with sanitizers, and we'd like to be aware when AFL toolchain does not
support a particular compilation flag. Currently we can WITH_ASAN=1,
but it silently drops sanitizer flags instead of complaining, which
gives an impression that we use sanitizer when in fact we do not.

* Support AFL builds with ASAN

AFL does not support 64-bit builds so build 32-bit binaries instead.
Note that we have to do in two places: during original make invocation
to build fuzzer tools, and during recursive make invocation to build
Themis library with AFL instrumentation.

* Build embedded BoringSSL for AFL

Pull in Soter dependencies when building Themis so that it triggers
BoringSSL build when ENGINE is set.

* Fix afl-gcc support

It turns out that afl-gcc actually uses AFL_CC environment variable
(who would have though, eh?) And if we set it to afl-gcc then afl-gcc
tries to invoke afl-gcc in an infinite recursive loop. Avoid that by
unexporting the variable, keeping it only inside make.

* Document sanitizer caveats

* Exercise fuzzers during CI builds

Introduce a new "fuzz" job that builds, tests, and runs fuzzers.
We check that we can build all interesting configurations and run
the most interesting ones for a bit to ensure that there are no
obvious mistakes in the code. We'd like to run longer, but CircleCI
has around 20-minute limit which does not fit what we want to eat.

AFL is meant to be used interactively (primarily) but it is possible
to script its execution too. We use a bit peculiar spell for that:

    timeout -s INT 30s make fuzz | cat -u || true

with the following rationale:

- timeout limits execution time of "make fuzz" which is by default
  unbounded. Put a 30-second limit on execution to keep it civil.
  We've got quite a few fuzzing tools to run.

- AFL expects to be cleanly stopped by SIGINT. However, it will stop
  with non-zero exit code, so we eat it with "|| true"

- Piping into cat disconnects AFL from terminal so it won't be
  screaming about "terminal window is too small, plz resize it
  to at least 80x25, kthx"). Everything is better with cats.

* Disable backtrace analysis on CI

CircleCI runs stuff in unprivileged Docker containers and gdb requires
ptrace capabilities to attach to processes. Unfortunately, we'll get
no stack traces on CI.

Well, add an option to skip debugger analysis and just try crashing,
maybe that will output something interesting.

* Improve error reporting in tools

Often fuzzing finds some length field, inverts it, and this results in
a gigabyte-sized allocation that malloc() refuses to fulfill. However,
it does not set "errno" which we use for error reporting at call site
resulting in silly errors: "failed to read file.dat: Success". Let's
set errno to something sensible before exiting.

Also, limit the input buffer sizes to 50 MB. AFL can bit-flip those
lengths too which is not what we really want. Since we use 32-bit
lengths those are mostly fine on 64-bit systems, but when running
32-bit binaries with ASAN malloc() tends to panic when allocating
more than 2 GB. Don't let it do that and abort the test (cleanly).

* Dump input file in crash reports

We don't have direct access to reports on CI so let's dump the input
into logs so that it's readily accessible. It's usually something
small so it's okay to write it out.

* Correctly report findings

It is possible for report directory to contain multiple reports, with
some not having crashes. "print_banner" gets reset for each report set
so we cannot use. Introduce a dedicated flag to note that we have had
a crash and should report that.

* Build 32-binaries only for ASAN + AFL

We'd been building 32-bit binaries when ASAN was enabled even withou
AFL. ASAN works fine on general 64-bit systems, we need to build 32-bit
binaries only when running AFL.

Use -m32 when building fuzzing tools. Variable adjuments when
AFL_USE_ASAN=1 will take care of the modified Themis build.

* Store full AFL reports on CircleCI

In addition to dumping information to log, save full reports in CircleCI
artifacts for each build so that they may be retrieved later. There are
quite a few small files in there so zip them into a single archive.