Explicitly disable Mac Catalyst support (#598)
Some of our targets declare that they support both iOS and iPadOS.
Themis has never supported iPadOS explicitly, so disable all that stuff.
Actually, instead of explicitly disabling, just set it to default (which
will be iOS only).

Recent versions of Xcode automatically enable "Mac Catalyst" -- support
layer allowing to run iPadOS applications on macOS. This breaks build
because OpenSSL and BoringSSL dependencies explicitly disable Catalyst.
To be on the safe side, explicitly disable Catalyst for all our projects.
We may enable it later, together with iPadOS support.

All of this is currently invisible to Bitrise because it builds with
older Xcode 10.2. However, macOS 10.14+ with Xcode 11.2 is affected
by this when building from command line.