Fix podspec versioning and tests (#599)
Remove the hack that we have hastily added to work around a packaging
issue with GRKOpenSSL 1.0.2.19. We still pin the version to 1.0.2.18
though, because the library has never been updated and 1.0.2.19 is still
broken as it was. However, we can't release the next version of Themis
pinned to 0.12.1. Use the tagged version. This also lets the podspec
to validate without warnings (from our part, at least).

Also remove the Podfile.lock from our unit test project. It is meant to
be used with the latest version from master as evident from the Podfile.
However, presence of Podfile.lock makes "pod install" to install an
older version of Themis 0.11 which is lacks some new features that are
being tested. (Bitrise is fine because it removes the lockfile manually.)

Remove the lock file to ensure that we are testing the trunk version
of Themis, and add Podfile.lock to .gitignore to prevent it from being
committed again on accident.