Check C compiler flags with AFL_CC when available
"supported" function is used to determine whether a flag is supported by
the compiler. Use AFL_CC instead of regular CC if available to correctly
check for flags when building AFL stuff.

Some time ago "supported" supported a second argument to select the
compiler to use, but this option does not seem to be used anymore.