Use emcmake when building BoringSSL with Emscripten
Emscripten toolchain includes "emconfigure" tool which sets various
standard variables to setup crosscompilation. Since v1.14.1 Emscripten
include "emcmake" tool which is effectively emconfigure for CMake.

Themis can build embedded BoringSSL when it is selected as cryptographic
backend. With Emscriptedn we always use (and build) embedded BoringSSL.
For that we use "emconfigure cmake" to configure the build.

Since Emscripten v1.39.9 it is now mandatory to use "emcmake" for CMake
configuration, as it is evident from the following error message:

    $ emmake make wasmthemis
    make: make wasmthemis
    building embedded BoringSSL...
    error: use `emcmake` rather then `emconfigure` for cmake projects
    make: *** [build/boringssl/crypto/libcrypto.a] Error 1

Use emmake if it is available, otherwise fall back to emconfigure when
building embedded BoringSSL with Emscripten.