Make string_to_pointer_size pure
This function is used to convert a string into an FFI::MemoryPointer
with string's bytes that we can pass to Themis Core. It's current
implementation is... not ideal because it uses "force_encoding".

This is problematic because it does not return a copy of a string with
different encoding. It modifies the encoding of the string in-place.
This is unexpected for client code *and* does not work for fronzen
strings (e.g., literal constants).

Update the implementation to use FFI::MemoryPointer.from_string which
has the same behavior (returns a pointer into string's buffer) but
does not change the encoding of the string.

This function is used throughout RbThemis code, not only by Secure Cell.
It should be enough to add encoding tests for Secure Cell only, for now.