Update Secure Cell API: Seal mode
Start preparing Secure Cell to passphrase API introduction by updating
encryption/decryption API. It is not directly related to passphrases,
but since we're going to reeducate users about Secure Cell, it's a good
point to improve the API.

One annoyance with (all) Secure Cell APIs is how they are mapped into
Swift. This is somewhat asymmetrical mapping: wrap and unwrapData,
because of some Swift renaming magic. This naming scheme is also
inconsistent with every other Themis wrapper on desktop systems which
all use encrypt end decrypt.

Update the API to use "encrypt" and "decrypt" naming scheme. Also,
provide more overloads for Objective-C to omit optional arguments.
Note that we have to hide some API from Swift in order to get better
error handling conversion. We need to expose only methods with explicit
NSError out-parameters.

The old API is declated deprecated from now on. It is implemented via
the new API as a compatibility shim.

While we're here, take this chance to improve inline API docs which make
for a better IDE experience.