Update Secure Cell API: Context Imprint mode
Context Imprint mode needs the same consistency updates as Seal mode.
However, things are much easier because associated context is not
optional in this mode.

If there is anything interesting, note that it's no longer necessary to
explicitly spell out __autoreleasing for NSError out parameters. Modern
Objective-C compilers are able to infer this annotation.